+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - live transpilation on lib-tests (#2163)
 - live transpilation on lib-common (#2164)
 - Add a widget provider for the classroom creation form
+- Allow delete playlist resources
+  - FileDepository 
 - Add classroom widgets :
   - InfoBar
   - Description